Missing Types Decorator policy ensures that each policy contains a Type
